Agrega los archivos a la carperta src
